Add rough new sub-clock update implementation

+++ AWO/Src/Game/Clock/Sub_Clock/sub_clock.h
@@ -1,7 +1,6 @@
 #pragma once
 
 #include "Game/Data/Clock/clock_data.h"
-#include "Game/Clock/Tick_Event/tick_event.h"
 #include "Game/Clock/Pub_Sub/_pub_sub.h"
 
 /*! @brief A sub-clock working under the game clock. Registers to the clock pub-sub service and 
@@ -22,7 +21,7 @@ Sub_Clock* create_sub_clock(Clock_Data* clock_data, Animation_Clock_Index index)
  *  @param[in] clock The sub-clock to update.
  *  @param[in] passed_frames The amount of game frames that passed since the last update.
  */
-void update_sub_clock(Sub_Clock* clock, unsigned int passed_frames);
+void update_sub_clock(Sub_Clock* clock, Uint8 passed_frames);
 
 /*! @brief Frees a sub-clock module.
  *
